Added two Reset method overloads. Added Local-Tests task.
GitReset(ResetMode) which resets the current branch HEAD to the latest
commit and, optionally, modifies the index and working tree.

GitReset(ResetMode, CommitId) which resets the current branch HEAD to the
specified commit and, optionally, modifies the index and working tree.

$Target can be passed a value of Local-Tests which runs all tests except
Git-Clone. The purpose of this is to allow for the situation where
someone is behind a corporate proxy and the test repository on GitHub
cannot be cloned (see https://github.com/libgit2/libgit2/issues/2106).