Add option to load tag targets before disposing the repository
When using GitTags, the libgit2sharp repository is disposed when the
tags are returned. Hence, accessing a tag's target properties, which are
lazily loaded, throws an exception.

Added an optional parameter (loadTargets) to GitTags. When set to true,
the target of all the tags are loaded before the repository is disposed,
enabling access to those. The loadTargets parameter defaults to false,
to make this an opt-in feature.

Fixes #72